+++ Core/Components/Transforms/ResolutionChangeEvt.cs
@@ -1,0 +1,9 @@
+using Unity.Entities;
+
+namespace UGUIDots.Transforms {
+
+    /// <summary>
+    /// Marks a messaging entity that the resolution has changed.
+    /// </summary>
+    public struct ResolutionChangeEvt : IComponentData { }
+}