+++ Core/_Prototype/ConstantScreenSizeLogger.cs
@@ -1,8 +1,0 @@
-﻿using UnityEngine;
-
-public class ConstantScreenSizeLogger : MonoBehaviour {
-
-    void Update() {
-        Debug.Log($"[Screen] Width: {Screen.width}, Height: {Screen.height}");
-    }
-}